requirements.txt
fastapi==0.95.1
uvicorn[standard]==0.22.0
celery==5.3.0
redis==4.5.5        # Python клиент для Redis (брокер и результат)
PyMuPDF==1.22.0     # fitz
pdfplumber==0.9.0
camelot-py==0.10.1
openpyxl==3.0.10    # (Camelot может требовать для эксель, хотя не обязательно)
pytesseract==0.3.10
Pillow==9.4.0
